requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
pillow==11.3.0
pymongo==4.13.2
sqlparse==0.5.3
tzdata==2025.2
venv\Scripts\activate